refactor(FE): 검색 쿼리 타입을 search.ts에서 관리하도록 변경 (#137)
