+++ client/src/types/search.ts
@@ -2,3 +2,13 @@ export interface Label {
   type: string;
   value: string;
 }
+
+export interface SearchQuery {
+  lastId?: string;
+  tags?: string[];
+  authors?: string[];
+  category?: string;
+  reviews?: number;
+  likes?: number;
+  detail?: string;
+}
